[REFACTOR] split gamemodes into destinct files
