+++ inexor/gamemode/hideandseek_common.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#define TEAM_HIDE "hide"
+#define TEAM_SEEK "seek"
+
+
+struct hideandseekmode
+{
+
+};
